Fix bugs in tail and append methods
The commit fixes bugs related to working with null instead of a node.
For append, I added a conditional for attempting to append
to an empty list. For tail, the fix had to do with checking if nextNode
is null rather than waiting for the while loop to hit the null node.